+++ CHANGELOG.md
@@ -3,5 +3,5 @@ All notable changes to the "cryptoticker" extension will be documented in this f
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-- Initial release
+## [This version]
+- Adding support for selecting an exchange.